Add files via upload
Posted in order to demonstrate the steep learning curve that comes with taking on practical projects. The code here implements various methods of data analysis in order to understand the relative importance of each feature, here I try various algorithms in order to achieve the most effective model. This data analysis allowed me to gain the insights that I apply onto the next and final model.